- Updated comment to reflect emailer tool - Fixed Pep 8 issues - Renamed file to be more consistent
